+++ core/scripts/version-check.py
@@ -49,6 +49,7 @@ RAVEN_RELEASES = [
     "5.5.3",
     "5.5.4",
     "5.5.5",
+    "5.5.6",
 ]
 
 RAVEN_LATEST = RAVEN_RELEASES[-1]